refactor: extract models into separate module to address circular imports
